# Performance improvement in AssertionInfo.