@horenmar 2017-06-22
Ill take a look at this on the train, but first one more question: what exactly was your benchmark and is it measured with any optimizations?