@coombez 2017-06-22
I can't provide the code, but basically -O3 optimized on gcc.  Pseudo code for the test with an rdtsc time source.

Time start = now();
for(int i = 0; i < 10000000; ++i)
    CHECK(now() - start < 10_seconds);

The time in here is just used as work to do that can't be completely optimized away, not for benchmarking.  For the runtime of the test, my build system tells me how long it ran, /usr/bin/time would work just fine for this purpose.  Or you could just output your equivalent of now() - start after the loop.