@horenmar 2017-06-22
Ok, we have something similarly simple as Benchmark binary...

I will merge this and open a new branch you should check out, it removes the second argument and thus quite a lot of the extra concatenation. It doesn't quite pass approval tests, but it would be fairly simple to fix that and the perf gain might be worthwhile.