@horenmar 2017-06-22
Ok, performance gains from `dev-second-string-argument-removal` should be final for those changes.