@coombez 2017-06-23
I tested out dev-second-string-argument-removal, it neither improved or regressed performance, so seems fine to me.

I've also discovered that the single most expensive operation remaining is the construction/destruction of the std::ostringstream.  removing this drops a test with 100M instead of 10M iterations from 60s->5.5s.  A further drop from 5.5s->0.2s is achievable by avoiding building the assertion results and 'handling' them in the case of success.  That said, I have not achieved either of these changes while maintaining your tests of course.  But it's still useful information for where to focus future changes.