@philsquared 2017-06-23
Yeah, I started a proof-of-concept rewrite of they core that did no eager conversions (and didn't use stringstreams at all.
I'm going to be taking a serious look at incorporating some of the ideas once I get v2.0 released.