@coombez 2017-06-23
PR #939 eliminates the construction/destruction of the std::ostringstream while passing the tests.  Hopefully another easy win.