@coombez 2017-06-28
For Catch v2.0, if you'll accept the suggestion, it may be worth benchmarking the assertions compared to the cost of an if-statement.  If that seems too restrictive, I wouldn't throw away that benchmark, but perhaps add a couple, one that include incrementing a count, not sure what the third should be but I'm sure you could come up with one.

Fundamentally, the real value Catch will add will be in the exceptional case of failure (or the user wanting full results), and that's where all the cost should lie.  Successful tests that are never reported should be as close to free as possible.

Once my last PR is merged, Catch assertions will be 40x faster! :)  But still 50x slower than a simple if.  Most of this is in the ostringstream.  But more than I expected is in the evaluation of whether successful tests should be included, in the increment of passed tests, and in the deferred checking of success or failure.

In Catch, I will reach a point of diminishing returns in continuing optimization soon, but if you're writing something from scratch, there are often ways to implement the features that will give you the benefits you want without costing in performance.  It's much harder to perfectly match results when optimizing after the fact, at least for large structural optimizations.