@philsquared 2017-06-28
For Catch 2 the plan is to defer (almost) all the cost to the reporters - by passing just references to the original, typed, objects that are still on the stack in the assertion. It's up the reporters whether they will obtain a stringified representation or not (depending on whether the assertion passed or not, whether the `-s` option has been passed, or whether the reporter always logs all results, regardless.

I did it this way in my proof-of-concept spike to Catch2 and it worked very well - so I'm going to try to steer the current implementation that way.